Assignment9Part2
I did make the changes to the index.html file in the href to reflect the relative path to the files. I also re-ran the ng build --prod command in terminal to see if the files all got included. I did create the employee-detail.html file in the app folder as directed in the instructions and copied the contents of the employee-detail.html file that was in the par 10 folder to that file and saved it. I do think I remember getting an error that referenced the employee-detail.html file when I ran the ng build --prod command in terminal before. So, I hope this cleared that up?!